+++ timeline/tests/test_ui.py
@@ -3,6 +3,7 @@
 import json
 
 from django.urls import reverse
+from django.utils.timezone import localtime
 from django.test import override_settings
 from urllib.parse import urlencode
 
@@ -15,6 +16,7 @@ from projectroles.models import SODAR_CONSTANTS
 from projectroles.plugins import get_backend_api
 from projectroles.tests.test_ui import UITestBase
 
+from timeline.models import TL_STATUS_OK
 from timeline.tests.test_models import (
     TimelineEventMixin,
     TimelineEventStatusMixin,
@@ -42,7 +44,6 @@ class TestProjectListView(
     def setUp(self):
         super().setUp()
         self.timeline = get_backend_api('timeline_backend')
-
         # Init default event
         self.event = self.timeline.add_event(
             project=self.project,
@@ -53,7 +54,6 @@ class TestProjectListView(
             extra_data=EXTRA_DATA,
             status_type=self.timeline.TL_STATUS_OK,
         )
-
         # Init classified event
         self.classified_event = self.timeline.add_event(
             project=self.project,
@@ -159,7 +159,6 @@ class TestSiteListView(
     def setUp(self):
         super().setUp()
         self.timeline = get_backend_api('timeline_backend')
-
         # Init default event
         self.event = self.timeline.add_event(
             project=None,
@@ -170,7 +169,6 @@ class TestSiteListView(
             extra_data=EXTRA_DATA,
             status_type=self.timeline.TL_STATUS_OK,
         )
-
         # Init classified event
         self.classified_event = self.timeline.add_event(
             project=None,
@@ -200,11 +198,10 @@ class TestSiteListView(
 
     def test_render_object(self):
         """Test visibility of object related events in site-wide event list"""
-        # Add user as an object reference
-        self.obj_ref = self.event.add_object(
+        obj_ref = self.event.add_object(
             obj=self.superuser, label='user', name=self.superuser.username
         )
-        self.classified_obj_ref = self.classified_event.add_object(
+        self.classified_event.add_object(
             obj=self.superuser, label='user', name=self.superuser.username
         )
         expected = [
@@ -221,32 +218,30 @@ class TestSiteListView(
         url = reverse(
             'timeline:list_object_site',
             kwargs={
-                'object_model': self.obj_ref.object_model,
-                'object_uuid': self.obj_ref.object_uuid,
+                'object_model': obj_ref.object_model,
+                'object_uuid': obj_ref.object_uuid,
             },
         )
         self.assert_element_count(expected, url, 'sodar-tl-list-event')
 
-    def test_object_button(self):
-        """Test visibility of the return button in event's object view"""
-        # Add user as an object reference
-        self.obj_ref = self.event.add_object(
+    def test_render_back_button(self):
+        """Test visibility of back button in object view"""
+        obj_ref = self.event.add_object(
             obj=self.superuser, label='user', name=self.superuser.username
         )
         url = reverse(
             'timeline:list_object_site',
             kwargs={
-                'object_model': self.obj_ref.object_model,
-                'object_uuid': self.obj_ref.object_uuid,
+                'object_model': obj_ref.object_model,
+                'object_uuid': obj_ref.object_uuid,
             },
         )
         self.login_and_redirect(
             self.superuser, url, wait_elem=None, wait_loc='ID'
         )
-        try:
-            self.selenium.find_element(By.CLASS_NAME, 'btn-secondary')
-        except Exception:
-            pass
+        self.assertIsNotNone(
+            self.selenium.find_element(By.ID, 'sodar-tl-back-btn')
+        )
 
 
 class TestAdminListView(
@@ -308,7 +303,6 @@ class TestModals(TimelineEventMixin, TimelineEventStatusMixin, UITestBase):
     def setUp(self):
         super().setUp()
         self.timeline = get_backend_api('timeline_backend')
-
         # Init default event
         self.event = self.timeline.add_event(
             project=self.project,
@@ -319,50 +313,89 @@ class TestModals(TimelineEventMixin, TimelineEventStatusMixin, UITestBase):
             extra_data=json.dumps(EXTRA_DATA),
             status_type=self.timeline.TL_STATUS_OK,
         )
-        self.event_status_init = self.event.set_status(
-            self.timeline.TL_STATUS_INIT,
-            'Event initialized',
-            extra_data={'example_data': 'example_extra_data'},
+        self.url = reverse(
+            'timeline:list_project', kwargs={'project': self.project.sodar_uuid}
         )
 
     def test_details_modal(self):
         """Test details modal"""
-        url = reverse(
-            'timeline:list_project', kwargs={'project': self.project.sodar_uuid}
-        )
         self.login_and_redirect(
-            self.superuser, url, wait_elem=None, wait_loc='ID'
+            self.superuser, self.url, wait_elem=None, wait_loc='ID'
+        )
+        link = self.selenium.find_element(By.CLASS_NAME, 'sodar-tl-link-detail')
+        link.click()
+        WebDriverWait(self.selenium, 10).until(
+            EC.presence_of_element_located((By.ID, 'sodar-tl-table-detail'))
         )
         self.assertIsNotNone(
-            self.selenium.find_element(By.CLASS_NAME, 'sodar-tl-link-detail')
+            self.selenium.find_element(By.CLASS_NAME, 'sodar-card-table')
+        )
+        title = self.selenium.find_element(By.CLASS_NAME, 'modal-title')
+        self.assertIn('Event Details: ', title.text)
+        elem = self.selenium.find_element(
+            By.CLASS_NAME, 'sodar-tl-event-detail-timestamp'
+        )
+        self.assertEqual(
+            elem.text,
+            localtime(self.event.get_timestamp()).strftime('%Y-%m-%d %H:%M:%S'),
+        )
+        elem = self.selenium.find_element(
+            By.CLASS_NAME, 'sodar-tl-event-detail-description'
+        )
+        self.assertEqual(elem.text, self.event.get_status().description)
+        elem = self.selenium.find_element(
+            By.CLASS_NAME, 'sodar-tl-event-detail-status-type'
+        )
+        self.assertEqual(elem.text, TL_STATUS_OK)
+
+    def test_details_modal_description_line_break(self):
+        """Test details modal with line breaks in description"""
+        desc = 'Test\nwith\nline breaks'
+        status = self.event.get_status()
+        status.description = desc
+        status.save()
+        self.login_and_redirect(
+            self.superuser, self.url, wait_elem=None, wait_loc='ID'
+        )
+        link = self.selenium.find_element(By.CLASS_NAME, 'sodar-tl-link-detail')
+        link.click()
+        WebDriverWait(self.selenium, 10).until(
+            EC.presence_of_element_located((By.ID, 'sodar-tl-table-detail'))
+        )
+        elem = self.selenium.find_element(
+            By.CLASS_NAME, 'sodar-tl-event-detail-description'
+        )
+        self.assertEqual(
+            elem.get_attribute('innerHTML'), desc.replace('\n', '<br>')
         )
 
-    def test_extra_modal(self):
+    def test_extra_data_modal(self):
         """Test extra data modal"""
-        url = reverse(
-            'timeline:list_project', kwargs={'project': self.project.sodar_uuid}
-        )
         self.login_and_redirect(
-            self.superuser, url, wait_elem=None, wait_loc='ID'
+            self.superuser, self.url, wait_elem=None, wait_loc='ID'
         )
-        self.assertIsNotNone(
-            self.selenium.find_element(
-                By.CLASS_NAME, 'sodar-tl-link-extra-data'
-            )
+        link = self.selenium.find_element(
+            By.CLASS_NAME, 'sodar-tl-link-extra-data'
         )
+        link.click()
+        WebDriverWait(self.selenium, 10).until(
+            EC.presence_of_element_located((By.CLASS_NAME, 'sodar-tl-json'))
+        )
+        title = self.selenium.find_element(By.CLASS_NAME, 'modal-title')
+        self.assertIn('Extra', title.text)
+        body = self.selenium.find_element(By.CLASS_NAME, 'modal-body')
+        self.assertIn('"{"test_key": "test_val"}"', body.text)
 
-    def test_status_extra_modal(self):
+    def test_status_extra_data_modal(self):
         """Test status extra data modal"""
-        url = reverse(
-            'timeline:list_project', kwargs={'project': self.project.sodar_uuid}
-        )
+        status = self.event.get_status()
+        status.extra_data = {'test': 170}
+        status.save()
         self.login_and_redirect(
-            self.superuser, url, wait_elem=None, wait_loc='ID'
+            self.superuser, self.url, wait_elem=None, wait_loc='ID'
         )
-        button = self.selenium.find_element(
-            By.CLASS_NAME, 'sodar-tl-link-detail'
-        )
-        button.click()
+        link = self.selenium.find_element(By.CLASS_NAME, 'sodar-tl-link-detail')
+        link.click()
         WebDriverWait(self.selenium, 10).until(
             EC.presence_of_element_located(
                 (By.CLASS_NAME, 'sodar-tl-link-status-extra-data')
@@ -374,73 +407,20 @@ class TestModals(TimelineEventMixin, TimelineEventStatusMixin, UITestBase):
             )
         )
 
-    def test_details_content(self):
-        """Test details modal content"""
-        url = reverse(
-            'timeline:list_project', kwargs={'project': self.project.sodar_uuid}
-        )
-        self.login_and_redirect(
-            self.superuser, url, wait_elem=None, wait_loc='ID'
-        )
-        button = self.selenium.find_element(
-            By.CLASS_NAME, 'sodar-tl-link-detail'
-        )
-        button.click()
-        WebDriverWait(self.selenium, 10).until(
-            EC.presence_of_element_located((By.ID, 'sodar-tl-table-detail'))
-        )
-        self.assertIsNotNone(
-            self.selenium.find_element(By.CLASS_NAME, 'sodar-card-table')
-        )
-        title = self.selenium.find_element(By.CLASS_NAME, 'modal-title')
-        self.assertIn('Event Details: ', title.text)
-        table = self.selenium.find_element(By.ID, 'sodar-tl-table-detail')
-        expected = [
-            'Timestamp',
-            'Description',
-            'Status',
-            self.timeline.TL_STATUS_INIT,
-        ]
-        for e in expected:
-            self.assertIn(e, table.text)
-
-    def test_extra_content(self):
-        """Test extra data modal content"""
-        url = reverse(
-            'timeline:list_project', kwargs={'project': self.project.sodar_uuid}
-        )
-        self.login_and_redirect(
-            self.superuser, url, wait_elem=None, wait_loc='ID'
-        )
-        button = self.selenium.find_element(
-            By.CLASS_NAME, 'sodar-tl-link-extra-data'
-        )
-        button.click()
-        WebDriverWait(self.selenium, 10).until(
-            EC.presence_of_element_located((By.CLASS_NAME, 'sodar-tl-json'))
-        )
-        title = self.selenium.find_element(By.CLASS_NAME, 'modal-title')
-        self.assertIn('Extra', title.text)
-        body = self.selenium.find_element(By.CLASS_NAME, 'modal-body')
-        self.assertIn('"{"test_key": "test_val"}"', body.text)
-
     def test_copy_button(self):
         """Test for copy button"""
-        url = reverse(
-            'timeline:list_project', kwargs={'project': self.project.sodar_uuid}
-        )
         self.login_and_redirect(
-            self.superuser, url, wait_elem=None, wait_loc='ID'
+            self.superuser, self.url, wait_elem=None, wait_loc='ID'
         )
-        button = self.selenium.find_element(
+        link = self.selenium.find_element(
             By.CLASS_NAME, 'sodar-tl-link-extra-data'
         )
-        button.click()
+        link.click()
         WebDriverWait(self.selenium, 10).until(
             EC.presence_of_element_located((By.CLASS_NAME, 'sodar-tl-copy-btn'))
         )
-        btn = self.selenium.find_element(By.CLASS_NAME, 'sodar-tl-copy-btn')
-        btn.click()
+        button = self.selenium.find_element(By.CLASS_NAME, 'sodar-tl-copy-btn')
+        button.click()
 
 
 class TestSearch(TimelineEventMixin, TimelineEventStatusMixin, UITestBase):
@@ -449,7 +429,6 @@ class TestSearch(TimelineEventMixin, TimelineEventStatusMixin, UITestBase):
     def setUp(self):
         super().setUp()
         self.timeline = get_backend_api('timeline_backend')
-
         # Init default event
         self.event = self.timeline.add_event(
             project=self.project,
@@ -460,14 +439,12 @@ class TestSearch(TimelineEventMixin, TimelineEventStatusMixin, UITestBase):
             extra_data=EXTRA_DATA,
             status_type=self.timeline.TL_STATUS_OK,
         )
-
         self.make_event_status(
             event=self.event,
             status_type=self.timeline.TL_STATUS_SUBMIT,
             description=self.timeline.TL_STATUS_SUBMIT,
             extra_data=EXTRA_DATA,
         )
-
         # Init default site event
         self.site_event = self.timeline.add_event(
             project=None,
@@ -478,14 +455,12 @@ class TestSearch(TimelineEventMixin, TimelineEventStatusMixin, UITestBase):
             extra_data=EXTRA_DATA,
             status_type=self.timeline.TL_STATUS_OK,
         )
-
         self.make_event_status(
             event=self.site_event,
             status_type=self.timeline.TL_STATUS_SUBMIT,
             description=self.timeline.TL_STATUS_SUBMIT,
             extra_data=EXTRA_DATA,
         )
-
         # Init classified event
         self.classified_event = self.timeline.add_event(
             project=None,
@@ -496,14 +471,12 @@ class TestSearch(TimelineEventMixin, TimelineEventStatusMixin, UITestBase):
             extra_data=EXTRA_DATA,
             classified=True,
         )
-
         self.make_event_status(
             event=self.classified_event,
             status_type=self.timeline.TL_STATUS_SUBMIT,
             description=self.timeline.TL_STATUS_SUBMIT,
             extra_data=EXTRA_DATA,
         )
-
         self.classified_site_event = self.timeline.add_event(
             project=None,
             app_name=APP_NAME_PR,
@@ -513,7 +486,6 @@ class TestSearch(TimelineEventMixin, TimelineEventStatusMixin, UITestBase):
             extra_data=EXTRA_DATA,
             classified=True,
         )
-
         self.make_event_status(
             event=self.classified_site_event,
             status_type=self.timeline.TL_STATUS_SUBMIT,