+++ CHANGELOG.rst
@@ -20,6 +20,8 @@ Changed
 - **Projectroles**
     - Update project role list inactive user rendering (#1658)
     - Move role list item template to ``_project_role_item.html`` (#1664)
+- **Timeline**
+    - Update event details modal rendering to support line breaks (#1660)
 
 
 v1.1.4 (2025-04-04)